fix: display Tomcat metrics and service statuses
