+++ backend/routes/serviceControl.js
@@ -1,20 +1,87 @@
 require('dotenv').config({ path: __dirname + '/.env' });
 
 const express = require('express');
+const os = require('os');
+const axios = require('axios');
+const { parseStringPromise } = require('xml2js');
 const router = express.Router();
+const { exec } = require('child_process');
 const path = require('path');
 const fs = require('fs').promises;
-const { getTomcatMetrics } = require('../services/tomcatMetrics');
 
 
 
 let TOMCAT_PROCESS_NAME = 'Tomcat*'; // default fallback
 
-
 router.get('/tomcat/metrics', async (req, res) => {
   try {
-    const metrics = await getTomcatMetrics();
-    res.json(metrics);
+
+    //const auth = { username: 'admin', password: 'admin' };
+    const TOMCAT_USERNAME = process.env.TOMCAT_USERNAME || 'admin';
+    const TOMCAT_PASSWORD = process.env.TOMCAT_PASSWORD || 'admin';
+    const auth = { username: TOMCAT_USERNAME, password: TOMCAT_PASSWORD };
+
+    //const tomcatBaseUrl = `${BACKEND_HOST}:${PORT}`;
+    const tomcatBaseUrl = `http://127.0.0.1:8080`;
+    
+    // Fetch XML status data
+    //const { data: xml } = await axios.get('http://localhost:8080/manager/status/all?XML=true', { auth });
+    const { data: xml } = await axios.get(`${tomcatBaseUrl}/manager/status/all?XML=true`, { auth });
+    const result = await parseStringPromise(xml, { explicitArray: false, mergeAttrs: true });
+    const status = result.status;
+    const jvm = status.jvm;
+    const connector = status.connector;
+
+    const memoryPools = Array.isArray(jvm.memorypool) ? jvm.memorypool : [jvm.memorypool];
+    const heapPools = memoryPools.filter(pool => pool.type === 'Heap memory');
+    const nonHeapPools = memoryPools.filter(pool => pool.type === 'Non-heap memory');
+
+    // Fetch deployed applications from text endpoint
+    //const { data: appListText } = await axios.get('http://localhost:8080/manager/text/list', { auth });
+    const { data: appListText } = await axios.get(`${tomcatBaseUrl}/manager/text/list`, { auth });
+    const appLines = appListText.split('\n').filter(line => line.startsWith('/'));
+    const applications = appLines.map(line => {
+      const [path, state, sessions, name] = line.trim().split(':');
+      return {
+        name: name || path.replace('/', '') || 'ROOT',
+        contextPath: path,
+        status: state.charAt(0).toUpperCase() + state.slice(1),
+        sessions: parseInt(sessions)
+      };
+    });
+
+    const server = await getServerInfo();
+
+    // Final JSON response
+    res.json({
+      server,
+      threads: {
+        max: parseInt(connector.threadInfo.maxThreads),
+        current: parseInt(connector.threadInfo.currentThreadCount),
+        busy: parseInt(connector.threadInfo.currentThreadsBusy),
+        utilization: (
+          (parseInt(connector.threadInfo.currentThreadsBusy) /
+            parseInt(connector.threadInfo.maxThreads)) *
+          100
+        ).toFixed(1)
+      },
+      requests: {
+        count: parseInt(connector.requestInfo.requestCount),
+        errors: parseInt(connector.requestInfo.errorCount),
+        avgProcessingTime:
+          parseInt(connector.requestInfo.processingTime / connector.requestInfo.requestCount) || 0,
+        timeout: 20000
+      },
+      memory: {
+        heap: calculateMemory(heapPools),
+        nonHeap: calculateMemory(nonHeapPools),
+        gc: {
+          count: 1247, 
+          time: 2345   
+        }
+      },
+      applications
+    });
   } catch (error) {
     res.status(500).json({
       error: 'Failed to fetch Tomcat metrics',
@@ -23,6 +90,83 @@ router.get('/tomcat/metrics', async (req, res) => {
   }
 });
 
+async function getServerInfo() {
+  return new Promise((resolve) => {
+    exec(`powershell -Command \"(Get-Process -Name '${TOMCAT_PROCESS_NAME}' | Select-Object -First 1).StartTime.ToString('yyyy-MM-dd HH:mm:ss')\"`, async (err, stdout) => {
+      const startTime = stdout.trim() || 'Unknown';
+      const uptime = startTime !== 'Unknown' ? calculateUptime(startTime) : 'Unknown';
+
+      resolve({
+        status: 'Running',
+        uptime,
+        startTime,
+        jvmVersion: await getJvmVersion(),
+        tomcatVersion: await getTomcatVersion(),
+        os: `${os.type()} ${os.release()}`
+      });
+    });
+  });
+}
+
+function calculateUptime(startTimeStr) {
+  const startTime = new Date(startTimeStr);
+  const now = new Date();
+  const uptimeMs = now - startTime;
+  const days = Math.floor(uptimeMs / (1000 * 60 * 60 * 24));
+  const hours = Math.floor((uptimeMs / (1000 * 60 * 60)) % 24);
+  const minutes = Math.floor((uptimeMs / (1000 * 60)) % 60);
+  return `${days} days, ${hours} hours, ${minutes} minutes`;
+}
+
+async function getJvmVersion() {
+  return new Promise((resolve) => {
+    exec('java -version', (err, stdout, stderr) => {
+      const match = stderr.match(/version \"(.*?)\"/);
+      resolve(match ? `OpenJDK ${match[1]}` : 'Unknown');
+    });
+  });
+}
+
+/*async function getTomcatVersion() {
+  return new Promise((resolve) => {
+    exec(
+      `powershell -Command "(Get-Content -Path 'C:\\Program Files\\Apache Software Foundation\\Tomcat 9.0_Tomcat9_64bit\\RELEASE-NOTES' -Raw) | Select-String -Pattern 'Apache Tomcat'"`,
+      (err, stdout) => {
+        const match = stdout.match(/Apache Tomcat.*?(\d+\.\d+\.\d+)/);
+        resolve(match ? `Apache Tomcat/${match[1]}` : 'Apache Tomcat/Unknown');
+      }
+    );
+  });
+}*/
+
+async function getTomcatVersion() {
+  const TOMCAT_BASE = process.env.TOMCAT_BASE || 'C:\\Program Files\\Apache Software Foundation\\Tomcat 9.0_Tomcat9_64bit';
+  const releaseNotesPath = path.join(TOMCAT_BASE, 'RELEASE-NOTES');
+  return new Promise((resolve) => {
+    exec(
+      `powershell -Command "(Get-Content -Path '${releaseNotesPath}' -Raw) | Select-String -Pattern 'Apache Tomcat'"`,
+      (err, stdout) => {
+        const match = stdout.match(/Apache Tomcat.*?(\d+\.\d+\.\d+)/);
+        resolve(match ? `Apache Tomcat/${match[1]}` : 'Apache Tomcat/Unknown');
+      }
+    );
+  });
+}
+
+
+function calculateMemory(pools) {
+  const used = pools.reduce((sum, pool) => sum + parseInt(pool.usageUsed || 0), 0);
+  const max = pools.reduce((sum, pool) => {
+    const val = parseInt(pool.usageMax);
+    return val > 0 ? sum + val : sum;
+  }, 0);
+
+  return {
+    usedMB: Math.round(used / 1024 / 1024),
+    maxMB: Math.round(max / 1024 / 1024),
+    usagePercent: max > 0 ? ((used / max) * 100).toFixed(1) : '--'
+  };
+}
 
 const logPositions = {};
