--- backend/services/tomcatMetrics.js
@@ -1,139 +1,0 @@
-const axios = require('axios');
-const { parseStringPromise } = require('xml2js');
-const os = require('os');
-const { exec } = require('child_process');
-const path = require('path');
-
-const TOMCAT_PROCESS_NAME = process.env.TOMCAT_PROCESS_NAME || 'Tomcat*';
-
-async function getTomcatMetrics() {
-  try {
-    const TOMCAT_USERNAME = process.env.TOMCAT_USERNAME || 'admin';
-    const TOMCAT_PASSWORD = process.env.TOMCAT_PASSWORD || 'admin';
-    const auth = { username: TOMCAT_USERNAME, password: TOMCAT_PASSWORD };
-    const tomcatBaseUrl = 'http://127.0.0.1:8080';
-
-    const { data: xml } = await axios.get(`${tomcatBaseUrl}/manager/status/all?XML=true`, { auth });
-    const result = await parseStringPromise(xml, { explicitArray: false, mergeAttrs: true });
-    const status = result.status;
-    const jvm = status.jvm;
-    const connector = status.connector;
-
-    const memoryPools = Array.isArray(jvm.memorypool) ? jvm.memorypool : [jvm.memorypool];
-    const heapPools = memoryPools.filter(pool => pool.type === 'Heap memory');
-    const nonHeapPools = memoryPools.filter(pool => pool.type === 'Non-heap memory');
-
-    const { data: appListText } = await axios.get(`${tomcatBaseUrl}/manager/text/list`, { auth });
-    const appLines = appListText.split('\n').filter(line => line.startsWith('/'));
-    const applications = appLines.map(line => {
-      const [ctx, state, sessions, name] = line.trim().split(':');
-      return {
-        name: name || ctx.replace('/', '') || 'ROOT',
-        contextPath: ctx,
-        status: state.charAt(0).toUpperCase() + state.slice(1),
-        sessions: parseInt(sessions)
-      };
-    });
-
-    const server = await getServerInfo();
-
-    return {
-      server,
-      threads: {
-        max: parseInt(connector.threadInfo.maxThreads),
-        current: parseInt(connector.threadInfo.currentThreadCount),
-        busy: parseInt(connector.threadInfo.currentThreadsBusy),
-        utilization: (
-          (parseInt(connector.threadInfo.currentThreadsBusy) /
-            parseInt(connector.threadInfo.maxThreads)) *
-          100
-        ).toFixed(1)
-      },
-      requests: {
-        count: parseInt(connector.requestInfo.requestCount),
-        errors: parseInt(connector.requestInfo.errorCount),
-        avgProcessingTime:
-          parseInt(connector.requestInfo.processingTime / connector.requestInfo.requestCount) || 0,
-        timeout: 20000
-      },
-      memory: {
-        heap: calculateMemory(heapPools),
-        nonHeap: calculateMemory(nonHeapPools),
-        gc: {
-          count: 1247,
-          time: 2345
-        }
-      },
-      applications
-    };
-  } catch (error) {
-    throw error;
-  }
-}
-
-async function getServerInfo() {
-  return new Promise((resolve) => {
-    exec(`powershell -Command \"(Get-Process -Name '${TOMCAT_PROCESS_NAME}' | Select-Object -First 1).StartTime.ToString('yyyy-MM-dd HH:mm:ss')\"`, async (err, stdout) => {
-      const startTime = stdout.trim() || 'Unknown';
-      const uptime = startTime !== 'Unknown' ? calculateUptime(startTime) : 'Unknown';
-
-      resolve({
-        status: 'Running',
-        uptime,
-        startTime,
-        jvmVersion: await getJvmVersion(),
-        tomcatVersion: await getTomcatVersion(),
-        os: `${os.type()} ${os.release()}`
-      });
-    });
-  });
-}
-
-function calculateUptime(startTimeStr) {
-  const startTime = new Date(startTimeStr);
-  const now = new Date();
-  const uptimeMs = now - startTime;
-  const days = Math.floor(uptimeMs / (1000 * 60 * 60 * 24));
-  const hours = Math.floor((uptimeMs / (1000 * 60 * 60)) % 24);
-  const minutes = Math.floor((uptimeMs / (1000 * 60)) % 60);
-  return `${days} days, ${hours} hours, ${minutes} minutes`;
-}
-
-async function getJvmVersion() {
-  return new Promise((resolve) => {
-    exec('java -version', (err, stdout, stderr) => {
-      const match = stderr.match(/version \"(.*?)\"/);
-      resolve(match ? `OpenJDK ${match[1]}` : 'Unknown');
-    });
-  });
-}
-
-async function getTomcatVersion() {
-  const TOMCAT_BASE = process.env.TOMCAT_BASE || 'C:\\Program Files\\Apache Software Foundation\\Tomcat 9.0_Tomcat9_64bit';
-  const releaseNotesPath = path.join(TOMCAT_BASE, 'RELEASE-NOTES');
-  return new Promise((resolve) => {
-    exec(
-      `powershell -Command \"(Get-Content -Path '${releaseNotesPath}' -Raw) | Select-String -Pattern 'Apache Tomcat'\"`,
-      (err, stdout) => {
-        const match = stdout.match(/Apache Tomcat.*?(\d+\.\d+\.\d+)/);
-        resolve(match ? `Apache Tomcat/${match[1]}` : 'Apache Tomcat/Unknown');
-      }
-    );
-  });
-}
-
-function calculateMemory(pools) {
-  const used = pools.reduce((sum, pool) => sum + parseInt(pool.usageUsed || 0), 0);
-  const max = pools.reduce((sum, pool) => {
-    const val = parseInt(pool.usageMax);
-    return val > 0 ? sum + val : sum;
-  }, 0);
-
-  return {
-    usedMB: Math.round(used / 1024 / 1024),
-    maxMB: Math.round(max / 1024 / 1024),
-    usagePercent: max > 0 ? ((used / max) * 100).toFixed(1) : '--'
-  };
-}
-
-module.exports = { getTomcatMetrics };
